feat: express.static config. fix: prices in prodcuts table to DOUBLE

+++ back-end/src/database/migrations/20230328192217-createProducts.js
@@ -14,7 +14,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.DECIMAL,
+        type: Sequelize.DOUBLE,
         allowNull: false,
       },
       urlImage: {
